backend/requirements.txt
# backend/requirements.txt

# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# LLM & AI
langchain==0.1.0
langchain-community==0.0.10
langchain-huggingface==0.0.20
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
chroma-db==0.4.8

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dateutil==2.8.2

# Data & Utils
requests==2.31.0
pillow==10.1.0
numpy==1.24.3
pandas==2.1.3

# Production
gunicorn==21.2.0
waitress==2.1.2

# Development
pytest==7.4.3
python-dotenv==1.0.0

# Arabic Support
arabic-reshaper==0.4.0
python-bidi==0.4.2